requirement.txt
absl-py==2.1.0
batchgenerators==0.25
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
einops==0.8.0
filelock==3.15.4
fsspec==2024.6.1
future==1.0.0
google-auth==2.34.0
google-auth-oauthlib==1.0.0
grpcio==1.66.1
huggingface-hub==0.24.6
idna==3.8
imageio==2.35.1
importlib_metadata==8.4.0
importlib_resources==6.4.4
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
linecache2==1.0.0
Markdown==3.7
MarkupSafe==2.1.5
monai==0.9.0
mpmath==1.3.0
networkx==3.1
nibabel==5.2.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
protobuf==5.28.0
ptflops==0.7.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.4
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sympy==1.13.2
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.9
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.66.5
traceback2==1.4.0
triton==2.1.0
typing_extensions==4.12.2
unittest2==1.1.0
urllib3==2.2.2
Werkzeug==3.0.4
zipp==3.20.1
